#7 refactor: 페이지 컴포넌트화2
