+++ client/src/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { RecoilRoot } from 'recoil';
-import './Styles/index.css';
 import App from './App';
 import GlobalStyle from './Styles/global';
